api/smartcontracts/voting: Add retire clause
Adds a new clause to the voting token contract that allows the administrator to send voting tokens that are marked as `closed` to `OP_RETURN`. Without this, there's no way to clean up voting token UTXOs from old meetings.

Because this changes the contract, this will require a blockchain reset of any stacks using the voting API.

Closes chain/chainprv#874.
Reviewers: @erykwalder