api/txdb: encapsulate storage in Store object
We used to inject a database handle into every public
function in txdb via the context. With this change, the
Store now holds an internal reference to the underlying
database, and it ignores whatever database may or may
not be in the context.

We also take this opportunity to introduce more explicit
dataflow. We use signatures to indicate which functions
require a database transaction, which functions require
a DB (to begin and commit transactions) and which
functions don't care.

This means that all public functions in package txdb
that need to access the database handle (all but one of
them) are now methods on the Store object. Any client
code that wants to use those methods needs an explicit
reference to the store.

This is the one step toward making our txdb storage
completely isolated from the manager/etc storage. We
still have some queries that directly access txdb tables
from other packages. Once that isolation is complete, it
will enable us to move txdb storage to different
underlying engines, such as an alternative Postgres
configuration optimized for immutable data, or leveldb,
or something else.

Closes chain/chainprv#898.
Reviewers: @erykwalder @bobg