api/issuer: simplify rules for updating ADP
If the user provides an asset definition in an issuance
input, we used to check if it was the same as the
existing definition and only update it if the new value
was different. Presumably this was meant as an
optimization. Instead, we now let the user decide
unilaterally whether they want to update the definition.
If the field is present in the input, we update it. This
behavior is simpler (easier to explain) and slightly
more powerful (the user can now update the definition
when the new value is the same as the existing value,
which was previously not possible).

This patch also removes no-longer-used function
DefinitionHashByAssetID in txdb.

Closes chain/chainprv#904.
Reviewers: @bobg