bin: avoid an extra command to edit commit message
Flag -e (or --edit) tells git commit to open the editor
even when -F is given. This makes landing a PR that much
faster. Every little bit helps.

Also, this avoids ending in an intermediate state where
the commit message contains lines beginning with '#'.
That would have happened when cfile contains such lines,
but the user aborted the commit in their editor.

Closes chain/chainprv#921.
Reviewers: @boymanjor