core/asset: denormalize "spent in pool" into account_utxos
This copies information contained in the pool_inputs
table into account_utxos. In the future, txdb.Store
won't store this information at all, that is,
pool_inputs will disappear. Now it's the responsibility
of the various other modules (manager, explorer, etc) to
track this information if they need to.

This implementaiton is a naive in-place update of a
field for each account utxo. We can certainly revisit
this approach if performance measurement reveals that
it's a problem.

Future work: update some tests to stop using txdb.Store.
Do similar work on other manager utxo tables (for the
smart contracts).

Closes chain/chainprv#945.
Reviewers: @bobg