core/smartcontracts/voting: redistribute clause
Allows reassigning voting tokens to multiple, other voting rights. This
means that it's now possible for there to be multiple utxos in the
voting token contract with the same voting right asset ID and the same
voting token asset ID.

Voting token utxos are now specified through the outpoint.

The pseudocode and the opcode script for the new contract are in this gist:
https://gist.github.com/jbowens/6d0f3ced71949a6aecabbdec1ac88b8a

Closes chain/chainprv#969.
Reviewers: @erykwalder