core/smartcontracts/voting: multi-party register
Alters the voting token contract Register clause to support splitting a
voting token utxo into many utxos during registration. Also, adds a
`registration_id` to the contract which _may_ be provided during
registration to associate the utxo lot with an external ID.

Custodians proxy voting can use the `registration_id` to supply the id
of the party on behalf of which they're voting.

https://gist.github.com/jbowens/4358d24e4ebdeef6abaaa229a728e77a

Closes chain/chainprv#971.
Reviewers: @erykwalder