core/smartcontracts/orderbook: remove FK from orderbook tables to utxos
In the future, txdb.Store will stop storing outputs in
a database table, so we cannot rely on a foreign key
referencing table utxos. This is the last dependency
from the order book logic on txdb.Store.

Future work: update all the order book tests to stop
using txdb.Store.

Closes chain/chainprv#970.
Reviewers: @bobg