cos/txscript: use embedded prevout data
Instead of using the prevout data loaded through a state.View, use the
prevout data embedded on tx inputs. This removes the cos/txscript's
package's dependency on the state.ViewReader interface.

As a temporary measure, this adds a `circulationFunc` type. Later
that will be replaced by a reference to the state tree or a new interface
that's backed by the state tree.

Closes chain/chainprv#992.
Reviewers: @kr