core/smartcontracts/voting: run tests in transactions
Using pgtest.NewTx is faster. On my machine, typical run
time for 'go test' in this package before the change:
13.261s. Typical run time after: 6.902s. Speedup is 1.92x.

One test in this package still uses pgtest.NewContext
(creating a whole new database for that test). In
principle, that test also should be able to use
pgtest.NewTx, but there is a bug that needs to be fixed
first. (It's not clear exactly where the bug is, in test
code or production code.) Meanwhile, we can speed up the
rest of the tests here.

Closes chain/chainprv#995.
Reviewers: @jbowens