cos: use state tree for bc utxo validation
When validating and generating blocks & txs, use the state tree instead
of loading blockchain prevouts from txdb. Pool prevouts are still loaded
from the database. As a consequence, txdb.view is no longer used and
is deleted in this PR.

Also, because circulation is read from the state tree, this commit
resolves chain/chainprv#916.

Closes chain/chainprv#1018.
Reviewers: @kr