core/explorer: refactor ListUTXOsByAsset
Refactors the explorer package's utxos indexing to always index new
utxos. If the historical utxos feature flag is enabled, it will update
spent outputs. If the feature flag is not enabled, it will delete
outputs as they're spent.

The explorer.ListUTXOsByAsset endpoint is refactored here to use the
historical_utxos table.

This commit will require a blockchain reset for anyone using
ListUTXOsByAsset.

Closes chain/chainprv#1023.
Reviewers: @bobg