docker: introducing chain-core developer edition
This provides a good base for a chain-core docker container.

```
$ cd $CHAIN
$ bin/build-ccde
```

The build command will create a docker image that includes
a build of cored and cored's latest schema. It will also
install postgres onto the docker image and initialize
the database with the schema. The cored api is exposed via 8080.

After you run the build-ccde command, you can boot the
new docker image.

Closes chain/chainprv#1025.
Reviewers: @kr