cos: delete Pool.GetPrevouts
Instead of having a separate Pool method for retrieving prevouts when
validating a tx to be inserted into the pool, this commit uses the
existing `Pool.GetTxs` method to retrieve the entire previous
transactions. This will probably have some minor performance
consequences until we start caching pool txs in memory.

This also means that `utxos_status` and `blocks_utxos` are no longer
used can be dropped.

Closes chain/chainprv#1028.
Reviewers: @kr