cmd/cored: Add build tag to disable https redirect
The secureheader plugin redirects requests made with http to the
equivalent https URL. This commit exposes an env variable to turn
that functionality off. The functionality will stay on by default. This
will be useful for chain core developer edition. Users will be able to
connect to a chain core without a needing a TLS cert.

Closes chain/chainprv#1035.
Reviewers: @kr