core/asset: save asset definitions outside txdb
Moves storage of asset definitions and asset definition pointers outside
the txdb package. The pointers are already contained within the state
tree, so their storage isn't important to blockchain validation.

Once the issuance totals are also moved outside of txdb, we should be
able to entirely remove the cos/state.View replacing it only with the
state tree.

Closes chain/chainprv#1054.
Reviewers: @kr