core: createAsset no longer returns assetIssuerID
This is a partial reversal of 841e1b2. We've decided to deprecate
the issuerNodeID field the createAsset response, rather than
replace it with assetIssuerID. It is somewhat redundant, since the
client provided the asset issuer ID to begin with, and is
inconsistent with appdb.AssetResponse (returned by getAsset),
which never included issuerNodeID in the first place.

Closes chain/chainprv#1056.
Reviewers: @erykwalder