cos: initialize FC in a safer order
Setting the L should happen immediately along with
allocating and initializing the rest of the FC object.

As a general rule of thumb, it's good to avoid calling
methods on an object (such as an FC instance) until
after it's completely initialized.

Also, it's a little easier to call LatestBlock on the
store directly here, and it makes it even more obvious
that we don't call methods on a not-fully-initialized
FC object.

We might also want to make FC.LatestBlock call
WaitForBlock; that would be broken without this change.

Closes chain/chainprv#1072.
Reviewers: @jbowens