cos/bc: Update sighash logic
Replace `SigHashCache` with a new full-service hash-caching object,
`SigHasher`.  Update it to match the latest spec - almost.  It still
uses `TxData.LockTime`, which is slated to be replaced with a pair of
MinTime/MaxTime fields.

The "want" hashes in `TestTxHashForSig` were chosen to make the tests
pass and have not been independently vetted for correctness, since
they will only change again soon anyway (when the MinTime/MaxTime
change happens, and other serialization-format changes stabilize).

Closes chain/chainprv#1078.
Reviewers: @kr