core/appdb: decouple users from projects
This commit separates user creation from projects and vice versa.
Invitations are no longer scoped to projects. A user's role is now
kept in the users table as a required field. Project creation is not
linked to a user and the members table has been dropped.

This commit is the last in a two part project to remove project-level
data hiding from the core. All data is available to all users within
a core now. Admin privileges include: creating invitations and
updating user roles.

Future work will include adding functionality to delete a core user.

Closes chain/chainprv#1091.
Reviewers: @ryandotsmith