cos/validation: allow wildcard issuances for now
Until clients have an opportunity to switch over to providing
the asset amount on issue tx sources, allow zero-amount issuances.
If the amount is provided in all of an assets issuances, the
input/output parity will be enforced.

Also, fixes a bug where issuance totals would not be updated if
the asset has previously been issued but had not been issued in
the previous block.

Closes chain/chainprv#1099.
Reviewers: @kr