core/leader: new package
This is a mild refactoring, just moving some code
around. Package cmd/cored is getting unwieldy, so we
split some of it into a new package dedicated to
determining the leader process.

This also fixes a latency bug when fetching new blocks.
Previously the loop (shared between generator and
fetcher) would wait up to the block generation period
between attempts to fetch a new block. Now it fetches
blocks immediately.

Closes chain/chainprv#1105.
Reviewers: @bobg