bin/git-land: use GitHub API for merging
The GitHub API is now capable of squashing and rebasing
pull request branches automatically. It also enforces
status checks, so it will refuse to merge a PR with
failing checks. This means we can now use GitHub to do
these tasks instead of reimplementing them ourselves.

Closes chain/chainprv#1117.
Reviewers: @jeffomatic



The GitHub API is now capable of squashing and rebasing
pull request branches automatically. It also enforces
status checks, so it will refuse to merge a PR with
failing checks. This means we can now use GitHub to do
these tasks instead of reimplementing them ourselves.

This makes git land both simpler and, as a nice side
effect, faster.