database/pg/pgtest: replace NewContext with NewDB
NewContext put the database handle inside a context
object "for convenience". However, it's bad, as a
general practice, to put persistent state such as a db
handle in a context, so we'll factor this behavior out
to the handful of call sites that need it. (And we will
continue to work on removing those remaining places that
still do it.)

While we're at it, we add two other things to the
signature of NewDB: the schema path for initialization
and the generated DB URL. We could have added another
entry point that leaves out those parameters, for the
convenience of call sites that don't need them, but
there aren't really enough of those to justify it (nor
should we encourage their proliferation).

As of this change, no package outside of pgtest uses the
Open function, so it should probably become unexported
in the near future.

Closes chain/chainprv#1124.
Reviewers: @jbowens