cos: remove `conflicts` BlockCallback parameter
The `conflicts` parameter is mostly unused and isn't strictly necessary.
Supporting it would also complicate crash recovery, because we won't
have a historical snapshot of the pool when the block landed.

The only current use of the parameter was in `account_utxos` for
populating the `spent_in_pool` column correctly. This commit drops the
`spent_in_pool` column and uses the `reservation_id` instead. The two
are identical except that only checking `reservation_id` will not detect
when an output is spent in a pool tx from another core. We expect this to
be rare, and something we might need to solve for more generally (not
just for account_utxos) in the future.

Closes chain/chainprv#1125.
Reviewers: @kr