cos: stateless validation when adding to pool
Removes stateful validation from fc.AddTx() when new txs are added to
the pool. This removes the need for non-leader hosts within a Core to
have an up-to-date state tree.

This isn't a dramatic change, because we already allow conflicting txs
into the pool if they spend the same pool prevout.

Closes chain/chainprv#1128.
Reviewers: @kr