cos: Store exposes LatestStateTree
Refactors cos.Store to expose

  LatestStateTree(context.Context) (*patricia.Tree, uint64, error)

instead of

  StateTree(context.Context, uint64) (*patricia.Tree, error)

This is a prerequisite to refactoring LatestBlock into
GetBlock(n) so that a Core knows where to start from on
start up.

In a later PR, the state tree will be stored on the FC
instead of fetched from the database every time we need it.

Closes chain/chainprv#1134.
Reviewers: @kr