core/query/cql: new package
Adds new package for the Chain Query Language (CQL).

Package cql evaluates expressions in the Chain Query Language (CQL).
A CQL query is a boolean expression with zero or more placeholder
values ($1, $2, etc) that are initially unconstrained. The query is
evaluated in an environment (such as a transaction object or a UTXO)
that determines the value of all non-placeholder terms. The query
and its fixed values together constrain the placeholders. Function
Eval finds the set of all placeholder values that satisfy the query.

Closes chain/chainprv#1159.
Reviewers: @kr