core/accounts: move utxo indexing logic to package
This commit moves the account utxo indexing logic from core/asset to
core/accounts. This does require an additional package to register
callbacks to the cos.FC, but it consolidates the account logic to one
package.

This commit also moves the implementations for txbuilder interfaces
Receiver and Reserver to the core/accounts package.

Closes chain/chainprv#1171.
Reviewers: @kr