core/api-spec: change "destroy" to "retire"
After much consideration (specifically around reactions from clients), 
I think it is best to use the term "retire" for removing assets 
from circulation. Time and again, the term "destroy" raises questions 
as to whether it removes the asset completely from the blockchain, 
including all record of past activity. However, when I explain 
it as "retire", it is always clear and understandable from the 
start that the asset is not longer in circulation, but its legacy remains.

And I'm sorry, Mr. Jones, it's time.

Closes chain/chainprv#1218.
Reviewers: @ryandotsmith