core/api-spec.md: simplify the SDK-Core interface
This change should not affect the user experience. It is
purely a matter of communication between SDK and cored.

The "retire" SDK method is simply a control program that
starts with a `RETURN` instruction. To keep the
interface narrow -- so it contains fewer primitives --
the SDK can generate such a control program immediately
when the user requests to "retire" units of an asset.

The implementation in SDKs, despite some conceptual
repetition, is exptected to be less code overall than
handling an extra action type in cored.

Closes chain/chainprv#1235.
Reviewers: @ryandotsmith