core/txbuilder: introduce actions
Previously, the txbuilder package took separate "Source" and
"Destination" types, which had interface properties "Reserver" and
"Receiver". A reserver was responsible for selecting utxos for inputs,
as well as producing any change destinations. Receivers created the
control programs that lived in the transaction's outputs.

This introduces a new concept, "Action" which can produce transaction
inputs and the instructions to sign them, as well as transaction
outputs. Some actions will produce only inputs, and some only outputs.

On the API side, actions are deserialized in a two-step process. The
first process identifies the type of action, and the second deserializes
into that action.

Closes chain/chainprv#1231.
Reviewers: @kr