sdk/java: refactor BaseQuery abstraction
The `BaseQuery` class has been renamed `BaseQueryBuilder`
and it's `QueryPointer` member is now a `Query`. A `Query`
represents a search request to the core and includes a ChQL
query and its parameters (as well as information about paging).
This change alleviates the confusion caused by overuse of the
word query, while formalizing the process of building a ChQL
query on the client.

Closes chain/chainprv#1310.
Reviewers: @ryandotsmith