core/query: query outputs by txid
We want list-output queries to be able to search by
outpoint, including the txid, and for the results to
include txid in the annotated output object. But we
don't want the full annotated transaction (in
list-transactions) to include the txid in all of the
transaction's outputs, just once at the top level.

Closes chain/chainprv#1358.
Reviewers: @ryandotsmith