cos: add blockchain crash recovery
Adds a new FC.Recover() function for recovering blockchain state
from a restart or crash. The resulting blockchain should be fully
valid and complete.

Every time a process is elected as a Core's leader, it runs
FC.Recover() to restore the state of the blockchain.

As written, FC.Recover() can perform significant amounts of
unnecessary work. In future commits, we can optimize it to be
more intelligent about which work it needs to redo.

Note that this commit does not include generator-specific crash
recovery (saving generated blocks before signature).

Closes chain/chainprv#1144.
Reviewers: @kr