cos/vm: allow 0/0 multisig as in spec
It is useful to support the degenerate case of m/n
multisig where m and n are both zero, especially in
development and test, but also in real contracts that
might select among multiple multisig clauses. We still
forbid 0/n when n is positive. We intend to protect
against possible unintentional 0/0 usage in other layers
of Chain Core.

This change matches chain/cp1chain/chainprv#92 and chain/cp1chain/chainprv#91.

Closes chain/chainprv#1400.
Reviewers: @bobg