cos: remove UpsertGenesisBlock
This change requires an operator to explicitly create an
initial block when setting up a new blockchain network. The
new block's consensus program contains the network's
block-signing keys and quorum size. It is allowed to provide
zero keys and a quorum size of zero; this is useful for
development and testing.

Closes chain/chainprv#1401.
Reviewers: @jbowens