dashboard: Support "filter" and "sum_by" parameters
Splits balance ChQL field into "Filter" and "Sum By" inputs, and removes 
references to ChQL throughout the app.

* The Balance "Sum By" field uses `asset_id` as both its placeholder text 
and default query value sent to the server. This prevents the user from 
ever seeing a balance such as:

    ```
{ 
      "amount": 500
}
```

    with no additional context

* The "X" button used to clear searches has been replaced by a button 
labeled "Reset" that appears below the "Search" button when a filter 
or sum_by parameter is submitted.

Closes chain/chainprv#1436.
Reviewers: @jeffomatic