core: enable submit and get-blocks RPCs on any core
We don't need to restrict these two core-core RPCs to
generator nodes. Submitting a transaction will be useful
when we want all nodes to pre-validate transactions. It is
also safe to fetch blocks from any node, not just the
generator. (That is the whole point of a blockchain, after
all!)

Future work: remove "/rpc" from the path as well. Really, no
need to distinguish between intra-node rpcs and client SDK
rpcs.

Closes chain/chainprv#1434.
Reviewers: @jbowens