core: test blockchain crash recovery
Adds a thorough test of blockchain crash recovery. It tests generating
and committing a block, simulating crashes at every database operation.
It then pg_dumps the database and ensures that each dump is identical.

Because of the nature of this test, it is _slow_ (~40 seconds). It will
not run unless you provide the `LONG=1` environment variable to
`go test`.

Closes chain/chainprv#1465.
Reviewers: @bobg