core/generator: collapse signers into one interface
This commit refactors the core/generator interface
to be narrower — to more precisely describe what
it requires, and include fewer incidental pieces that
are not true dependencies.

It replaces the two concrete types *RemoteSigner and
*blocksigner.Signer with a single BlockSigner interface.
It is now up to the implementation whether it signs
locally or delegates to a remote process.

This refactoring is meant to accomplish two things:

1. Make core/generator easier to test by removing
   hard dependencies on other core packages.

2. Make the main logic of getAndAddBlockSignatures
   easier to follow by collapsing local and remote
   signers into a single interface.

TestGetAndAddBlockSignatures now no longer uses
the database at all.

This change loses some performance optimizations. In
particular, it can serialize the block multiple times,
and it no longer attempts to avoid making any remote
calls in the case where a local signature is sufficient.
(But in that case, it still cancels any remote calls and
does not wait for them.) If measured performance turns
out to be a problem, we can reintroduce those
optimizations or equivalent ones in a way that still
preserves the nice abstraction at this layer.

Closes chain/chainprv#1378.
Reviewers: @jbowens