protocol: add WaitForBlockSoon
The existing WaitForBlock returned an error when the
requested height is in the distant future. For some
callers, this meant calling the function was more
cumbersome than necessary — they either needed to check
for an impossible error or explain in comments why the
error doesn't need to be checked. (Or both!)

Instead, we modify WaitForBlock to do exactly what it
says, and introduce a variation in case the caller does
want the extra "distant future" error check. This only
happens when waiting for a block height derived from
user input, and is there to prevent resource abuse, not
for semantic correctness.

While we're at it, we also remove the unused context
parameter.

Closes chain/chainprv#1508.
Reviewers: @jbowens