core: fix response flush before self-exec
If we close the connection before the client reads all
data, the client considers that an error. We can make
that less likely by sending less data. Package http
sends the whole header in a single packet, so in
practice it seems safe to exec (which causes the OS to
reset the connection) after successfully transmitting
the header. Tests with curl confirm this.

Closes chain/chainprv#1576.
Reviewers: @bobg