core: fix race between reset and save snapshot
It's possible for the validator to save a snapshot after
reset has truncated the tables. This puts the core in an
inconsistent state and was causing a crash in recovery.
There are other defensive measures we should apply in
recovery, but to start with, this change should suffice.

We reset after exec, rather than before. This ensures
that no other goroutines (such as a validator saving a
snapshot) are running concurrently with the reset
function.

Closes chain/chainprv#1592.
Reviewers: @jbowens