protocol/vm: allow unordered signatures in MULTISIG
Currently, the Core doesn't guarantee assembled signatures 
are in the same order as the public keys in the control program. 
We should fix that, but as a temporary solution don't require 
the ordering to be the same.

Also, use asset.Signer.Quorum for figuring out 
how many signatures to include when signing an issuance.

Closes chain/chainprv#1605.
Reviewers: @kr