protocol/prottest: add NewChainWithStorage
This gives us a convenient way to make a test Chain
while specifying a Store and Pool.

We also use this in chain/core/generator to speed up a
test. Updating this test required giving up on our
(racy) attempt to test the synchronization; now it just
tests that GetBlocks gets blocks that are already
available.

Closes chain/chainprv#1608.
Reviewers: @jbowens