core: inline call to InitializeSigningGenerator
This was the only place that called
InitializeSigningGenerator, and it was easier to replace
it with a call to prottest.NewChain plus separate calls
to the asset and account init functions.

This doesn't help performance at all, but it will make
future changes easier; specifically, we'll want to
replace asset.Init and account.Init with constructor
functions and get rid of their package-global
configuration.

Closes chain/chainprv#1618.
Reviewers: @jbowens