cmd/corectl: match dashboard config behavior
This new subcommand 'config-generator' matches the
dashboard's behavior when configuring a new core as a
generator. It calls the same function, core.Configure,
to write the config, so it will hopefully be easier to
keep these to in sync with each other.

We also deleted subcommand 'init'.

Future work: add subcommand 'config' for non-generator
cores.

Closes chain/chainprv#1625
Reviewers: @tessr