cmd/cored: automatically init schema in dev builds
This will let us produce a single, self-contained binary
that can operate as a full Chain Core, given an empty
postgres database (as created by createdb).

It was just as easy to have it also build this way for
local dev, so I went ahead and did it that way.

Closes chain/chainprv#1633
Reviewers: @erykwalder