protocol: store latest state on Chain
Store the current block and snapshot on the protocol.Chain struct.
The values are updated on commitBlock() and will only be correct
if the current process is the leader. Users can compare the block's
height to the state.height to verify if it's current.

This fixes the issue where a single-host Core setup as a remote
signer couldn't sign blocks because it tried to read the snapshot
from the database (which may or may not have an up-to-date snapshot).

Future work:
  * forward sign block requests to leader if not the leader

Closes chain/chainprv#1639
Reviewers: @kr