core/txbuilder: populate MaxTime in Build
When building a transaction, clamp its max time to five minutes from
now. This will make it easier to detect tx rejections. Also, clamp the max
time by the reservation TTL as well.

In the future, in the worst case, /submit will wait until it sees a
block with a timestamp higher than the tx's max time before concluding
that the tx was rejected.

Closes chain/chainprv#1649
Reviewers: @erykwalder