core: add xpubs and quorum to API spec
This spec revisions proposes new API behavior: returning an
explicit list of xpubs and a quorum value for asset objects. While
this data is theoretically derivable from the issuance program, it
is helpful and convenient for developers to see the same values
that were provided when the asset was created. This is also
consistent with the account response object, which includes these
values.

Closes chain/chainprv#1654
Reviewers: @erykwalder