cmd/benchcore: new command
This command spins up a few EC2 instances running postgres, cored, and a
client to measure performance. Currently the client is a hello-world curl
request. In the future, it will be a java program to generate meaningful
load and measure throughput, latency, etc.

This is just a start. There are many possible optimizations we can to do
make the setup process potentially faster. It currently takes several
minutes.

Closes chain/chainprv#1622
Reviewers: @bobg @erykwalder