cmd/cored: set Last-Modified HTTP header field for dashboard files
The http package also uses this argument to serve
conditional requests, such as If-Modified-Since, more
efficiently.

Of course, the process start time isn't when the files
were last modified, but it's guaranteed to be after the
true last-modified time, which is a conservative (i.e.
safe) upper bound for the purposes of caching.

Closes chain/chainprv#1665
Reviewers: @dominic