protocol/bc: JSON null should unmarshal to zero-value Hash and AssetID
Previously, unmarshaling null into bc.Hash and bc.AssetID would
yield errors, since their UnmarshalTexts method could not handle
non-string values. However, when using either of these types in
structs, it is useful to treat null as the zero value. This commit
makes null the equivalent of undefined (i.e., a missing field in the
JSON of the enclosing struct), which already unmarshals to the
zero value.

Closes chain/chainprv#1675
Reviewers: @erykwalder @kr