protocol/validation: more precise interface for overflow-check add
This patch changes two things about this function:
its name and its return type.

The name addCheckOverflow is more descriptive than
safeSum. The second return value really communicates a
single bit of information. (Note that before this patch,
we were discarding the value of err -- a sign that
either there was a bug at the call site, or the
function's signature could be refined. This is an
example of the second case.)

Closes chain/chainprv#1677
Reviewers: @erykwalder