core/api-spec: list-transactions should take order param
Presently, transactions are returned in newest-to-oldest order.
This commit proposes an update to allow for oldest-to-newest
ordering. This is a stepping stone for implementing the
notification feature, and is useful in general.

When implementing this change, we'll want to be careful to update
the way we handle the cursor parameter. In the case of ascending
order, the internal SQL query should to reverse the filter
operator (> instead of <) used with the cursor.

Closes chain/chainprv#1688
Reviewers: @kr