core/api-spec: more intuitive pagination field names
Per discussion with Devon, we'd like add directional inflection
to our pagination-related field names:

- The "query" field in a pagination response has been renamed to
"next", which is more suggestive of the object's use, and accords
better to API patterns in the outside world. It also future-proofs
us for the introduction of "prev" pagination, should we choose to
support it.
- The "cursor" parameter for our various list-object API calls has
been renamed to "after". "after" refers to pagination order, as
opposed to the lexicographic or ordinal value of the cursor itself.
It is less ambiguous than "cursor", and future-proofs us for the
introduction of "before", should we choose to support reverse
pagination.

These changes can be implemented with a blanket find/replace.

Closes chain/chainprv#1687
Reviewers: @tessr