core: update pagination field names
This brings core to parity with https://github.com/chain/chainprv/pull/1687. 

We still use the word "cursor" all over the internals of core; if we are
going to start using the word `cursor` to refer to a persistent object
(e.g. for notifications) I might try to remove that word as well. 

Closes chain/chainprv#1691
Reviewers: @kr