cmd/chainbot: rely on github approval
We no longer need to check for LGTM, because our code
review system (github) has this built in. Their API has
the ability to refuse to land a PR that hasn't been
approved.

Note: this also means the commit message itself no
longer contains the list of reviewers. The list is
stored on github in the pull request's history.

Closes chain/chainprv#1695
Reviewers: @erykwalder @ryandotsmith