core/api-spec: make account/asset key info more explicit
This commit proposes the following:

1. Annotate accounts with information about the actual xpub used
to derive control programs, and provide distinction between that
xpub and the one used to create (i.e., the one used with SDK
signer objects) the account.
2. Ditto, for assets.
3. Make the "root_xpub" language consistent; i.e. so it also
surfaces in the object creation flow. Note that this implies
changes to language in the SDKs and the dashboard.
4. Annotate external assets (those that weren't created locally)
with pubkeys and quorum info, derived from the issuance program.

Closes chain/chainprv#1717