sdk/java: remove qa directory in favor of maven integration tests
Previously, the qa directory was used to create an uber jar that could be
added to
testing environments containing the core source code. Adding the java sdk
to the
monorepo allows us to run integration tests from the same environment as
said
source, without having to build a jar. One-off java test programs can also
be kept in
the sdk's test directory and built from the command line.

This setup also allows for unit tests to be added without a core
dependency. The
test suites are run with different commands `mvn test` vs. `mvn
integration-test`.

Closes chain/chainprv#1757