cmd/testbot: update testbot to run inside docker container
Previously, the service ran on the host machine and
started a new container for each test run. This should
speed up each run because the maven dependencies
are only installed once (as opposed to on each container
startup). This update also reduces the host dependencies
for testbot. Only docker is required. 

Closes chain/chainprv#1820