dashboard: add suggestions to tx asset and account alias fields
When typing into the account or asset alias fields, a list of partially
matching aliases will be displayed. Alias can be selected with arrow keys +
enter or the mouse.

Upon first autocomplete attempt, all items of the requested object type are
eagerly loaded into the store to provide accurate results.

Closes chain/chainprv#1851