core/mockhsm: cache xprivs in memory
The profile of the IOU benchmark in chain/core showed a significant amount
of time (10.54%) in HSM.load. Caching the key value in memory reduces the
time to sign as expected. However, it had little impact on overall
throughput, suggesting that the load call is not inherently slow, but
rather that the high latency in that call comes from blocking on something
else.

    benchmark                    old ns/op     new ns/op     delta
    BenchmarkSign-8              243973        51486         -78.90%
    BenchmarkIOUThroughput-8     17966586      17483572      -2.69%

Note that this implementation never invalidates cache entries. If a user
deletes a key, it will persist in the cache and continue to be usable to
sign things.

Closes chain/chainprv#1880